Rename some test filenames by convention
