+++ tests/snippets/builtin_complex.py
@@ -130,3 +130,24 @@ assert '(1-1j)' == repr(1-1j)
 # __getnewargs__
 assert (3 + 5j).__getnewargs__() == (3.0, 5.0)
 assert (5j).__getnewargs__() == (0.0, 5.0)
+
+
+class Complex():
+    def __init__(self, real, imag):
+        self.real = real
+        self.imag = imag
+
+    def __repr__(self):
+        return "Com" + str((self.real, self.imag))
+
+    def __sub__(self, other):
+        return Complex(self.real - other, self.imag)
+
+    def __rsub__(self, other):
+        return Complex(other - self.real, -self.imag)
+
+    def __eq__(self, other):
+        return self.real == other.real and self.imag == other.imag
+
+assert Complex(4, 5) - 3 == Complex(1, 5)
+assert 7 - Complex(4, 5) == Complex(3, -5)
