+++ tests/snippets/builtin_exec.py
