+++ tests/snippets/stdlib_abc.py
