+++ tests/snippets/stdlib_array.py
