+++ tests/snippets/stdlib_collections.py
