+++ tests/snippets/stdlib_hashlib.py
