+++ tests/snippets/stdlib_imghdr.py
