+++ tests/snippets/stdlib_io_bytesio.py
