+++ tests/snippets/stdlib_io_stringio.py
