+++ tests/snippets/stdlib_math.py
