+++ tests/snippets/stdlib_re.py
