+++ tests/snippets/stdlib_string.py
