+++ tests/snippets/stdlib_struct.py
