+++ tests/snippets/stdlib_sys.py
