+++ tests/snippets/stdlib_time.py
